Partially implemented character loading/unloading
Characters now start "unloaded", meaning that their process and
physics_process steps get skipped over. Debug keys have been added
to instruct rooms to register which enemies are inside their borders, and
load and unload them. This functions are supposed to be called on
room load, but there's nothing guaranteeing the enemies inside the
room are loaded before the room itself. So, I'll probably need to take
loading into my own hands, and make a loading step that happens on
the first process frame, and only then will actors be registered and such.