Room Loading and Solving
Trapped rooms will now lock doors on entry. Rooms now have a solution
function. When all the enemies in a room are killed, all trapped doors will
open, and any chests in the room will become active. Enemies are
properly registered within a room's boundaries, and are loaded and
unloaded when the room is loaded and unloaded.