Edited subtitle, added preview pictures
The title was "The Myth of Primordial Judgment" when it was supposed to be "The Myth of Primordial Retribution"

I added some preview pictures that I posted to Twitter

Pressing any key on the title screen now loads the debug dungeon.

There is a problem when loading the debug dungeon, if you don't wait long enough after loading, and enter a trapped room, the room's trap will not activate properly for some reason.

Chests now show the item contained within above them when they open. However, the image just remains floating there forever, and there's no text prompt or sound effect.