Added New Enemy: InvestiGator
The InvestiGator lunges with a biting attack. Currently, it travels randomly like the Takodachi, and attacks when you press the jump button. This behavior is only for testing. Unlike the Takodachi, this enemy should have a better AI, and should actually attempt to pursue Watson, targeting her with its attacks.

It has facing sprites and attacking sprites, but still needs walking sprites. It also needs a contact hitbox (likely a small one), an attack hitbox (only active during a couple frames of the lunge-bite) and to have its hurtbox and collision resized appropriately (and assigned to the correct collision layers)