Demo Release, Save file display, Licensing
The demo has been "completed," and is now in the .exports/ directory.

File select screen has been disabled; the game loads the demo dungeon
after pressing a key on the title screen.

A rudimentary save file format exists, and its information can be
displayed on the file select. It doesn't properly read or write to the game, however.

The license for Justin Walsh's castledb plugin for GODOT is now included, as it should have been earlier. A license for this program is included now too.