Aerial Gunshot on Release. Reorganize Input Code.
While mid-air, the gun now fires when releasing the shoot button. The input code for the player character, Watson, has also been restructured to be more readable, and better commented.

Watson's inputs are now polled in physics_process, rather than called on the instant they're pressed. This allows more control over the strict processing order of these inputs. Running code instantly on a button press is risky anyways, because there's no limit on the number of times it can be run per frame.