Added Events
When the Game object is told to play an event, the game pauses player and enemy movement. The event can then perform its own actions in its own process function. Once the event is complete, the Game must acknowledge its completion, and unpause. Acknowledged events free themselves.

Text events must be dismissed with the jump key in order to be deemed complete. Currently, they are the only event with a function, and they simply print their text to the console.