Area Loading and ItemTextEvents
Save files now store a string that indicates the area scene to load Locally, I have 3 save files. One loads Misteavous Manor, one loads the Debug Dungeon, and the last one loads the event testing room.

Implemented ItemTextEvents. This is the first compound event. It creates an ItemEvent and a TextEvent. It then runs the ItemEvent, waits for it to finish, then runs the TextEvent, and waits for the player to dismiss it.