Added Textboxes to TextEvents
Text events will now display the entered text in an on-screen
textbox.

Pressing spacebar to dismiss a text event which is too long to be
displayed in its entirety will instead display the next 2 lines. Pressing spacebar when the text event's last line or 2 are
currently displayed will free the textbox and end the Text event.

The main HUD node is now the node that gets assigned to Game.hud at the start of the game booting, instead of the healthbar.