Made dungeons identify rooms with a number
Rooms now have a room_id, and the dungeon will store rooms in a dictionary, where their room_id is the key and the dungeon's path to the room is the value. Even if a room changes name or path, its id number will be set quite intentionally. This will allow better saving and loading of room states in save files, while allowing me to change details of the room, such as its name and its location in the hierarchy, and it will still load properly, so long as it has the right id.

If a room's id is set to the default value, -1, it will be assigned the next available number and added to the dictionary.