Added Basic Save/Load and Chest/Door IDs
Pressing "P" will save the game. As of yet, there is no check that a save file is even loaded. Normally this function won't be able to be triggered from anywhere, so this is only a problem for the moment.

Inventory items, the current dungeon map, and the state of that dungeon's doors, chests, and rooms, are all that is loaded from save files at the moment. There likely isn't sufficient checks for whether or not the save file has data on the current dungeon.

Chests, rooms, and doors are now all assigned a unique id integer in the editor, and that is how the save files determine which data corresponds to which chest/room/door.