SaveData Stores Info In Variables
SaveData now stores its information in variables, rather than relying on everything to read the raw dictionary data. SaveData can read the raw dictionary data and convert it into its own simple variable structures. It can also convert its own data back into a dictionary.

The goal is eventually to pass references to SaveData's properties to their respective in-game objects, and then whenever an in-game object's savable property is changed, it will be reflected in SaveData. So when saving, I won't have to collect changed info from the world at the moment of saving. As a result, changes to the world will make changes to the SaveData immediately, and they merely need to be serialized and written to a file.